+++ src/javascript/controls/tabs.js
@@ -244,6 +244,7 @@ module.exports = {
             //Any tab opened
             filepath = null;
             plang = "";
+            editor = null;
             graviton.refreshStatusBarLinesAndChars( working_tab.getAttribute("screen"))
             graviton.changeLanguageStatusBar(plang, working_tab.getAttribute("screen"))
             document.getElementById( working_tab.getAttribute("screen")).children[1].children[0].style = "visibility:visible; display:block;";